Moved test from JMS to FunctionTest
