+++ Tests/Functional/JMSFunctionalTest.php
@@ -114,52 +114,6 @@ class JMSFunctionalTest extends WebTestCase
         ], $this->getModel('VirtualProperty')->toArray());
     }
 
-    public function testSymfonyConstraintDocumentation()
-    {
-        $this->assertEquals([
-            'required' => [
-                'propertyNotBlank',
-                'propertyNotNull',
-                'propertyAssertLengthRequired',
-            ],
-            'properties' => [
-                'propertyNotBlank' => [
-                    'type' => 'integer',
-                ],
-                'propertyNotNull' => [
-                    'type' => 'integer',
-                ],
-                'propertyAssertLengthRequired' => [
-                    'type' => 'integer',
-                    'minLength' => '1',
-                ],
-                'propertyAssertLengthMinAndMax' => [
-                    'type' => 'integer',
-                    'maxLength' => '50',
-                    'minLength' => '0',
-                ],
-                'propertyRegex' => [
-                    'type' => 'integer',
-                    'pattern' => '.*[a-z]{2}.*',
-                ],
-                'propertyCount' => [
-                    'type' => 'integer',
-                    'maxItems' => '10',
-                    'minItems' => '0',
-                ],
-                'propertyChoice' => [
-                    'type' => 'integer',
-                    'enum' => ['choice1', 'choice2'],
-                ],
-                'propertyExpression' => [
-                    'type' => 'integer',
-                    'pattern' => 'If this is a tech post, the category should be either php or symfony!',
-                ],
-            ],
-            'type' => 'object',
-        ], $this->getModel('SymfonyConstraints')->toArray());
-    }
-
     protected static function createKernel(array $options = [])
     {
         return new TestKernel(true);
